winmm: Don't do anything in `WINMM_BeginPlaying` if the device is already playing
This stops wave output devices' callbacks from being called reentrantly when they call `waveOutWrite`.

It was occuring because `waveOutWrite` calls `WINMM_BeginPlaying`, which calls `WOD_PushData`, which checks for completed audio frames and calls the callback to notify it if there are any.

By only calling `WOD_PushData` if the device isn't already playing, this effectively eliminates the problem. The callback will only be called when a frame is completed, so for it to be called when the device is stopped, it must have been the completion of the last frame in the queue. However, that means that no more `WOM_DONE` events are occuring which could trigger the callback.

If it's called on the completion of the second-last frame, `device->stopped` can't be set by calling `waveOutWrite`, since it's updated within `WOD_PushData`, which won't get called if the device isn't stopped.

It is still possible to trigger the callback reentrantly by other means (namely, by using `waveOutRestart`), but the desired behaviour is only that it can't be called reentrantly through just `waveOutWrite`.